e2e/util.go: refactor DeleteCluster() for common use between different frameworks
