+++ test/e2e/cluster_status_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/coreos/etcd-operator/pkg/spec"
 	"github.com/coreos/etcd-operator/pkg/util/k8sutil"
 	"github.com/coreos/etcd-operator/pkg/util/retryutil"
 	"github.com/coreos/etcd-operator/test/e2e/e2eutil"
@@ -45,7 +46,7 @@ func testReadyMembersStatus(t *testing.T) {
 	}
 
 	defer func() {
-		if err := deleteEtcdCluster(t, f, testEtcd); err != nil {
+		if err := e2eutil.DeleteEtcdCluster(t, f.KubeClient, testEtcd, &e2eutil.StorageCheckerOptions{}); err != nil {
 			t.Fatal(err)
 		}
 	}()
@@ -84,7 +85,18 @@ func testBackupStatus(t *testing.T) {
 		t.Fatal(err)
 	}
 	defer func() {
-		if err := deleteEtcdCluster(t, f, testEtcd); err != nil {
+		var storageCheckerOptions *e2eutil.StorageCheckerOptions
+		switch testEtcd.Spec.Backup.StorageType {
+		case spec.BackupStorageTypePersistentVolume, spec.BackupStorageTypeDefault:
+			storageCheckerOptions = &e2eutil.StorageCheckerOptions{}
+		case spec.BackupStorageTypeS3:
+			storageCheckerOptions = &e2eutil.StorageCheckerOptions{
+				S3Cli:    f.S3Cli,
+				S3Bucket: f.S3Bucket,
+			}
+		}
+
+		if err := e2eutil.DeleteEtcdCluster(t, f.KubeClient, testEtcd, storageCheckerOptions); err != nil {
 			t.Fatal(err)
 		}
 	}()
