+++ test/e2e/e2eutil/tpr_util.go
@@ -1,3 +1,17 @@
+// Copyright 2017 The etcd-operator Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package e2eutil
 
 import (
@@ -9,8 +23,14 @@ import (
 	"github.com/coreos/etcd-operator/pkg/spec"
 	"github.com/coreos/etcd-operator/pkg/util/k8sutil"
 
+	"github.com/aws/aws-sdk-go/service/s3"
 	"k8s.io/client-go/kubernetes"
 )
+
+type StorageCheckerOptions struct {
+	S3Cli    *s3.S3
+	S3Bucket string
+}
 
 func CreateCluster(t *testing.T, kubeClient kubernetes.Interface, namespace string, cl *spec.Cluster) (*spec.Cluster, error) {
 	uri := fmt.Sprintf("/apis/%s/%s/namespaces/%s/clusters", spec.TPRGroup, spec.TPRVersion, namespace)
@@ -28,6 +48,23 @@ func CreateCluster(t *testing.T, kubeClient kubernetes.Interface, namespace stri
 
 func UpdateEtcdCluster(kubeClient kubernetes.Interface, cl *spec.Cluster, maxRetries int, updateFunc k8sutil.ClusterTPRUpdateFunc) (*spec.Cluster, error) {
 	return k8sutil.AtomicUpdateClusterTPRObject(kubeClient.CoreV1().RESTClient(), cl.Metadata.Name, cl.Metadata.Namespace, maxRetries, updateFunc)
+}
+
+func DeleteEtcdCluster(t *testing.T, kubeClient kubernetes.Interface, cl *spec.Cluster, storageCheckerOptions *StorageCheckerOptions) error {
+	uri := fmt.Sprintf("/apis/%s/%s/namespaces/%s/clusters/%s", spec.TPRGroup, spec.TPRVersion, cl.Metadata.Namespace, cl.Metadata.Name)
+	if _, err := kubeClient.CoreV1().RESTClient().Delete().RequestURI(uri).DoRaw(); err != nil {
+		return err
+	}
+	if err := waitResourcesDeleted(t, kubeClient, cl); err != nil {
+		return err
+	}
+	if cl.Spec.Backup != nil {
+		err := waitBackupDeleted(kubeClient, cl, storageCheckerOptions)
+		if err != nil {
+			return fmt.Errorf("fail to wait backup deleted: %v", err)
+		}
+	}
+	return nil
 }
 
 func logfWithTimestamp(t *testing.T, format string, args ...interface{}) {
